Task 3 finished
Added filter component which filters the contact list by the entered input.

Filter component contains DOM of of this element.

Edited Phonebook component to add handleFilterChange function which takes a value and saves it to state.
The filteredContacts function then filters the contacts properties in state and leaves only the results that match the filter property.
Also added initial contacts to state.